% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_daytime_class.R
\name{as.daytime}
\alias{as.daytime}
\alias{as.daytime.default}
\alias{as.daytime.POSIXt}
\alias{as.daytime.character}
\title{Retrieve daytime information from a timestamp}
\usage{
as.daytime(x, ...)

\method{as.daytime}{default}(x, ...)

\method{as.daytime}{POSIXt}(x, rational = FALSE, first_min = 0, ...)

\method{as.daytime}{character}(x, rational = FALSE, first_min = 0, ...)
}
\arguments{
\item{x}{a timestamp that inherits from \code{character} or \code{POSIXt}}

\item{...}{arguments passed to \code{as.POSIXct}}

\item{rational}{logical. Return partial minutes as a decimal addend?}

\item{first_min}{value for the first minute of the day. Must be either
\code{0} (default; minutes coded as 0-1439) or \code{1} (minutes coded as
1-1440)}
}
\value{
an object with class \code{daytime}
}
\description{
Retrieve daytime information from a timestamp
}
\note{
If \code{x} is not already POSIX-formatted, coercion will be attempted.
  Arguments can be passed to \code{as.POSIXct} to assist with coercion,
  notably the \code{format} and \code{tz} arguments. Presumably the timezone
  should not matter, except perhaps during a daylight saving transition hour.
  But it is always safest to pass in POSIX-formatted data that you have
  vetted yourself.
}
\examples{

Time <- "2000-01-01 00:01:30"
t1 <- as.POSIXct(Time, "UTC")
t2 <- as.POSIXct(Time, "EST")
as.daytime(Time)
as.daytime(t1, TRUE)
as.daytime(t2, TRUE, 1)

## Beware of silent timezone changes. System will
## assume your local timezone unless told otherwise!
as.daytime(
  strftime(t1, "\%H:\%M:\%S", "EST"),
  format = "\%H:\%M:\%S"
)

}
