% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_as_circular.R
\name{as_circular}
\alias{as_circular}
\title{Cast an object as circular}
\usage{
as_circular(x, ...)
}
\arguments{
\item{x}{object to cast}

\item{...}{arguments passed to \code{\link{as_daytime}}}
}
\description{
Cast an object as circular
}
\details{
For non-\code{daytime} objects, the default method is a chain of
  casts, first to \code{daytime}, then to \code{circular}. Extra arguments
  can be passed through \code{...} to assist with the initial conversion to
  \code{daytime}
}
\keyword{internal}
