% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_package_doc_global_vars_and_imports.R
\docType{package}
\name{daytime}
\alias{daytime}
\title{daytime: Operate on time variables for physical behavior research}
\description{
Crucial functions of the \code{daytime} package are to allow conversion of
timestamps to numeric time of day (e.g., \code{0} for midnight and
\code{1439} for 11:59) and subsequent calculation of circular mean and
standard deviation.
}
