% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytime_methods__main.R
\name{daytime_methods}
\alias{daytime_methods}
\alias{mean.daytime}
\alias{sd.daytime}
\title{Methods for class \code{daytime}}
\usage{
\method{mean}{daytime}(x, ...)

\method{sd}{daytime}(x, ...)
}
\arguments{
\item{x}{a daytime object}

\item{...}{arguments passed to methods (currently unused)}
}
\description{
Methods for class \code{daytime}
}
\examples{

Time <- as_daytime(
  Sys.time()+rnorm(100, 12*60*60, 12*60*60), TRUE
)

## Wrap in `as.numeric` for better printing
as.numeric(mean(Time))
as.numeric(sd(Time))

## Compare
mean(as.numeric(Time))
sd(as.numeric(Time))
PAutilities::mean_sd(Time)

}
