% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_minute.R
\name{get_minute}
\alias{get_minute}
\title{Convert timestamp to minute of day}
\usage{
get_minute(timestamp, rational = FALSE, first_min = 0, ...)
}
\arguments{
\item{timestamp}{timestamp input, can inherit from \code{POSIXt} or
\code{character}, and can be either scalar or vector}

\item{rational}{logical. Return partial minutes as a decimal addend?}

\item{first_min}{value for the first minute of the day. Must be either
\code{0} (default; minutes coded as 0-1439) or \code{1} (minutes coded as
1-1440)}

\item{...}{arguments passed to \code{as.POSIXct}}
}
\value{
Values formatted as minute of day
}
\description{
Convert timestamp to minute of day
}
\examples{
Time <- Sys.time()
Time
get_minute(Time)
get_minute(Time, TRUE)
get_minute(Time, TRUE, 1)
}
