% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytime_methods__main.R
\name{is.daytime}
\alias{is.daytime}
\title{Test if an object belongs to \code{daytime} class}
\usage{
is.daytime(x, ...)
}
\arguments{
\item{x}{object to test}

\item{...}{arguments passed to \code{\link{range_test}}}
}
\description{
Test if an object belongs to \code{daytime} class
}
\examples{
x <- as_daytime(Sys.time(), FALSE)
is.daytime(x)
}
