% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytime_methods__main.R
\name{plot.daytime}
\alias{plot.daytime}
\title{Plot a \code{daytime} object}
\usage{
\method{plot}{daytime}(x, ...)
}
\arguments{
\item{x}{a \code{daytime} object}

\item{...}{arguments passed to \code{\link{as_circular}}}
}
\description{
Plot a \code{daytime} object
}
\details{
\code{x} is first cast as \code{circular}, then forwarded to the
  plot method for \code{circular} objects

plot(as_daytime(
  seq(
    as.POSIXct(Sys.Date()),
    as.POSIXct(Sys.Date()+1),
    "3 hour"
   )
))
}
