% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_checks.R
\name{range_fun}
\alias{range_fun}
\alias{order_limits}
\alias{range_examine}
\alias{range_test}
\title{Test object ranges against expectations}
\usage{
range_fun(x, limit, include_limit, include_fun, exclude_fun)

order_limits(lower, upper)

range_examine(
  x,
  lower = 0,
  upper = 1440,
  inc_lower = TRUE,
  inc_upper = FALSE,
  test_limits = TRUE
)

range_test(
  x,
  lower = 0,
  upper = 1439,
  inc_lower = TRUE,
  inc_upper = FALSE,
  rational_adjust = TRUE
)
}
\arguments{
\item{x}{object to test}

\item{limit}{placeholder for \code{lower} or \code{upper}}

\item{include_fun}{function for comparing \code{x} against \code{limit} when
the limit is included in the range}

\item{exclude_fun}{function for comparing \code{x} against \code{limit} when
the limit is excluded from the range}

\item{lower}{lower limit of expectation}

\item{upper}{upper limit of expectation}

\item{inc_lower}{logical. Include \code{lower} in the range (\code{[}) or
not (\code{(})?}

\item{inc_upper}{logical. Include \code{upper} in the range (\code{]}) or not
(\code{)})?}

\item{test_limits}{logical. run \code{order_limits} to check that
\code{lower} is less than \code{upper}}

\item{rational_adjust}{logical. Adjust the range based on a \code{rational}
attribute in \code{x}?}
}
\description{
Test object ranges against expectations
}
\keyword{internal}
