% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytime_arithmetic.R
\name{subtract_daytime}
\alias{subtract_daytime}
\title{Perform subtraction on \code{daytime} objects}
\usage{
subtract_daytime(e1, e2)
}
\arguments{
\item{e1}{a POSIX-formatted \code{x} attribute from a \code{daytime} object}

\item{e2}{a POSIX-formatted \code{x} attribute from a \code{daytime} object}
}
\value{
a \code{difftime} object, in minutes
}
\description{
Perform subtraction on \code{daytime} objects
}
\details{
This is a decidedly internal function that is strictly a wrapper
  around \code{difftime}. The magic happens with \code{Ops} dispatch of the
  \code{`-`} operator, where the objects are checked and the original
  timestamps pulled out.
}
\examples{
base_time <- Sys.time()
t1 <- as_daytime(base_time)
t2 <- as_daytime(base_time + 86460)

t1 - t2
t2 - t1

}
\seealso{
\code{daytime:::Ops.daytime}
}
\keyword{internal}
