% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytime__class.R
\name{as_daytime}
\alias{as_daytime}
\alias{as_daytime.POSIXt}
\alias{as_daytime.numeric}
\alias{as_daytime.circular}
\alias{as_daytime.character}
\alias{as_daytime.integer}
\title{Retrieve minute of day}
\usage{
as_daytime(x, ...)

\method{as_daytime}{POSIXt}(x, rational = FALSE, ...)

\method{as_daytime}{numeric}(x, rational = attr(x, "rational"), ...)

\method{as_daytime}{circular}(x, rational = attr(x, "rational"), ...)

\method{as_daytime}{character}(x, rational = FALSE, ...)

\method{as_daytime}{integer}(x, ...)
}
\arguments{
\item{x}{value from which to determine the minute of day value; typically
inherits from \code{character} or \code{POSIXt}, but methods are available
for \code{circular} and \code{numeric} as well}

\item{...}{arguments passed to \code{as.POSIXct}. See details.}

\item{rational}{logical. If \code{FALSE}, values are rounded down to the
nearest whole minute; if \code{TRUE}, real-numbered values are returned
with the decimal portions reflecting seconds}
}
\value{
a \code{daytime} object
}
\description{
Retrieve minute of day
}
\details{
If \code{x} inherits from \code{character}, casting to \code{POSIXt}
  will be attempted. Arguments can be passed to \code{as.POSIXct} to assist
  with this, notably the \code{format} and \code{tz} arguments. Presumably
  the timezone should not matter, except perhaps during a daylight saving
  transition hour. But it is always safest to pass in POSIX-formatted data
  that you have vetted yourself.
}
\examples{

t1_char <- "00:01:30"
t1 <- as.POSIXct(t1_char, "UTC", format = "\%H:\%M:\%S")

## Methods exist to give comparable output from character and POSIX inputs
as_daytime(t1_char, format = "\%H:\%M:\%S")
as_daytime(t1)

as_daytime(t1, TRUE)

## Beware of silent timezone changes. System will
## assume your local timezone unless told otherwise!
as_daytime(
  strftime(t1, "\%H:\%M:\%S", "EST"),
  format = "\%H:\%M:\%S"
)

}
