% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytime_methods__main.R,
%   R/daytime_methods_helpers.R
\name{daytime_methods}
\alias{daytime_methods}
\alias{mean.daytime}
\alias{sd.daytime}
\alias{sd}
\title{Methods for class \code{daytime}}
\usage{
\method{mean}{daytime}(x, ...)

\method{sd}{daytime}(x, units = c("min", "hr"), type = c("MSD", "SRL", "circular"), ...)

sd(x, ...)
}
\arguments{
\item{x}{input data}

\item{...}{arguments passed to methods}

\item{units}{character scalar specifying the desired units of output for
\code{sd.daytime}. Can be either \code{"min"} (for minutes) or \code{"hr"}
(for hours)}

\item{type}{character scalar specifying the method to use for \code{sd}.
Choose from the mean shorter distance (the default; \code{type = "MSD"}),
scaled resultant length (\code{type = "SRL"}), or \code{type =
"circular"} to call \code{circular::sd.circular}}
}
\description{
Methods for class \code{daytime}
}
\examples{

Time <- as_daytime(
  Sys.time()+rnorm(100, 0, 4*60*60), TRUE
)

## Wrap in `as.numeric` for better printing
as.numeric(mean(Time))
as.numeric(sd(Time))

## Compare
mean(as.numeric(Time))
sd(as.numeric(Time))
if (!!requireNamespace("PAutilities", quietly = TRUE)) PAutilities::mean_sd(Time)

}
