% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_get_minute.R
\name{get_minute}
\alias{get_minute}
\title{Convert timestamp to minute of day}
\usage{
get_minute(x, rational = FALSE, first_min = 0, ...)
}
\arguments{
\item{x}{a timestamp that inherits from \code{character} or \code{POSIXt}}

\item{rational}{logical. Return partial minutes as a decimal addend (for
numeric operations) or second value (for \code{strftime} operations)?}

\item{first_min}{value for the first minute of the day. Must be either
\code{0} (default; minutes coded as 0-1439) or \code{1} (minutes coded as
1-1440)}

\item{...}{arguments passed to \code{as.POSIXct}}
}
\description{
This is a generic-like function that will extract the numeric minute of day
if \code{x} inherits from \code{character} or \code{POSIXt}, or will
perform the reverse operation (returning a string with \code{\%H:\%M:\%S}
format) if \code{x} inherits from \code{integer} or \code{numeric}.
}
\note{
The value of \code{first_min} is stored as an attribute of the output
  when \code{x} inherits from \code{character} or \code{POSIXt}. This value
  is referenced when attempting re-conversion, and a warning will be issued
  if there is a conflict between the attribute and what is passed by the
  user. (See the commented example below.)
}
\examples{

## Extracting the numeric minute
Time <- Sys.time()
Time
get_minute(Time)
get_minute(Time, TRUE)
get_minute(Time, TRUE, 1)

## Reversing the operation
daytime <- get_minute(Time, TRUE)
get_minute(daytime)
get_minute(daytime, TRUE)
# get_minute(daytime, TRUE, 1)
}
