% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tod.R
\name{tod}
\alias{tod}
\alias{tod.default}
\alias{tod.daytime}
\alias{tod.POSIXt}
\title{Retrieve time of day in \code{\%H:\%M:\%S} or \code{\%H:\%M:00} format}
\usage{
tod(x, ...)

\method{tod}{default}(x, rational = FALSE, ...)

\method{tod}{daytime}(x, rational = attr(x, "rational"), ...)

\method{tod}{POSIXt}(x, rational = FALSE, ...)
}
\arguments{
\item{x}{a \code{daytime} or \code{POSIXt} object, or an object that can be
cast to \code{daytime}}

\item{...}{arguments passed to \code{\link{as_daytime}}.}

\item{rational}{logical. Return partial minutes as a numeric \code{\%S}
value? The default (\code{rational = FALSE}) will return \code{\%H:\%M:00}
format.}
}
\value{
character value(s) in the desired format
}
\description{
Retrieve time of day in \code{\%H:\%M:\%S} or \code{\%H:\%M:00} format
}
\details{
For objects that do not inherit from \code{daytime} or
  \code{POSIXt}, the default method is a chain of casts, first to
  \code{daytime} and then to \code{character}.
}
\examples{
x <- Sys.time()
y <- as_daytime(x)

tod(x)
tod(y)
tod(x, rational = TRUE)

tod(0)
tod(720)
tod(1440.9, 1, TRUE)
}
